fix reseting cart quantity bug when reload.
Previously, add to cart will redict to cart with query.
Dispatch action happens in CartScreen according to query.
When update qty in cart and reload, cart will be reset to query's qty.

This bug is fixed by moving dispatch to where it should happen:
add to cart button in ProductDetailScreen.